requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.40
python-dotenv
langchain-huggingface
langchain-chroma
langchain-xai
langchain-text-splitters
sentence-transformers
chromadb
bs4